# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DYOTAK",
    "description": "Orbital intelligence for ground-level survival \u2014 Satellite flood-damage mapping system for rescue teams.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Submit Job",
        "description": "Submit a new pipeline run.",
        "operationId": "submit_job_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Status",
        "description": "Poll job progress, stage timing, and status.",
        "operationId": "get_job_status_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel Job",
        "description": "Cancel a running or queued job.",
        "operationId": "cancel_job_api_jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/result": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Result",
        "description": "Retrieve result manifest containing layer URLs and provenance.",
        "operationId": "get_job_result_api_jobs__job_id__result_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultManifest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/facts": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Facts",
        "description": "Retrieve single source of truth facts.json for this job.",
        "operationId": "get_job_facts_api_jobs__job_id__facts_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactsJSON"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jobs/{job_id}/report": {
      "get": {
        "tags": [
          "report"
        ],
        "summary": "Get Report",
        "description": "Generate or retrieve one-page situation report.",
        "operationId": "get_report_api_jobs__job_id__report_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(en|ne)$",
              "default": "en",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CutoffSettlementRecord": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Settlement name from OSM place/town tag"
          },
          "class": {
            "$ref": "#/components/schemas/SettlementClass",
            "description": "Cut-off classification"
          },
          "buildings_affected": {
            "type": "integer",
            "title": "Buildings Affected",
            "description": "Number of flooded/affected buildings in cluster"
          },
          "extra_distance_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Distance Km",
            "description": "Detour km if still connected, null if cut off"
          },
          "priority_rank": {
            "type": "integer",
            "title": "Priority Rank",
            "description": "Rescue priority rank (1 = highest priority)"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "class",
          "buildings_affected",
          "priority_rank"
        ],
        "title": "CutoffSettlementRecord",
        "description": "Entry in cutoff_settlements table."
      },
      "DegradationItem": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable degradation code"
          },
          "reason_key": {
            "type": "string",
            "title": "Reason Key",
            "description": "i18n catalogue key explaining degradation"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Additional context"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "reason_key"
        ],
        "title": "DegradationItem",
        "description": "Disclosed degradation entry."
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "AOI_TOO_LARGE",
          "NO_VALID_ORBIT_PAIR",
          "NO_POST_EVENT_SCENE",
          "CDSE_UNAVAILABLE",
          "OHSOME_UNAVAILABLE",
          "MODEL_LOW_CONFIDENCE",
          "S2_TOO_CLOUDY",
          "BUSY",
          "INTERNAL"
        ],
        "title": "ErrorCode",
        "description": "Canonical error codes defined in ARCHITECTURE.md."
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Machine-readable error code"
          },
          "message_key": {
            "type": "string",
            "title": "Message Key",
            "description": "Key for i18n message catalogue"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "description": "Template parameters for translation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message_key"
        ],
        "title": "ErrorDetail",
        "description": "Structured error payload carrying locale message_key and parameter dict."
      },
      "FactItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Machine-readable fact key e.g. flooded_area_km2"
          },
          "value": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "boolean"
              }
            ],
            "title": "Value",
            "description": "Fact value computed by pipeline"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "description": "Unit of measurement e.g. km2, count, km"
          },
          "source_stage": {
            "$ref": "#/components/schemas/PipelineStage",
            "description": "Pipeline stage producing this fact"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "value",
          "unit",
          "source_stage"
        ],
        "title": "FactItem",
        "description": "An individual atomic fact item."
      },
      "FactsJSON": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/FactsMeta"
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/FactItem"
            },
            "type": "array",
            "title": "Facts"
          },
          "tables": {
            "$ref": "#/components/schemas/FactsTables"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta",
          "facts",
          "tables"
        ],
        "title": "FactsJSON",
        "description": "Single source of truth numbers generated by Stage 8 (Facts)."
      },
      "FactsMeta": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "bbox": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Bbox"
          },
          "event_date": {
            "type": "string",
            "title": "Event Date"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "units": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Units"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "bbox",
          "event_date",
          "provenance"
        ],
        "title": "FactsMeta",
        "description": "Metadata enclosing facts.json."
      },
      "FactsTables": {
        "properties": {
          "cutoff_settlements": {
            "items": {
              "$ref": "#/components/schemas/CutoffSettlementRecord"
            },
            "type": "array",
            "title": "Cutoff Settlements"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "FactsTables",
        "description": "Structured tabular data in facts.json."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobOptions": {
        "properties": {
          "use_baseline": {
            "type": "boolean",
            "title": "Use Baseline",
            "description": "Force log-ratio baseline instead of ML model",
            "default": false
          },
          "optical_enabled": {
            "type": "boolean",
            "title": "Optical Enabled",
            "description": "Attempt Sentinel-2 optical cue extraction",
            "default": true
          },
          "max_pre_scenes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Pre Scenes",
            "description": "Max pre-event acquisitions to median"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "JobOptions",
        "description": "Optional execution tunables for a job."
      },
      "JobRequest": {
        "properties": {
          "bbox": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Bbox",
            "description": "Bounding box [min_lon, min_lat, max_lon, max_lat] in WGS84"
          },
          "event_date": {
            "type": "string",
            "title": "Event Date",
            "description": "Event date in YYYY-MM-DD format"
          },
          "osm_snapshot_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Osm Snapshot Date",
            "description": "Optional OSM snapshot date; must strictly precede event_date"
          },
          "options": {
            "$ref": "#/components/schemas/JobOptions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "bbox",
          "event_date"
        ],
        "title": "JobRequest",
        "description": "Payload to trigger a full flood and damage mapping run."
      },
      "JobState": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "JobState",
        "description": "High-level state of a pipeline job."
      },
      "JobStatus": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id",
            "description": "Unique job identifier"
          },
          "status": {
            "$ref": "#/components/schemas/JobState",
            "description": "Overall job state"
          },
          "current_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PipelineStage"
              },
              {
                "type": "null"
              }
            ],
            "description": "Currently active stage if running"
          },
          "progress": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "description": "Overall progress fraction [0.0, 1.0]",
            "default": 0.0
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageStatus"
            },
            "type": "array",
            "title": "Stages",
            "description": "Per-stage progress records"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO 8601 job creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "ISO 8601 last update timestamp"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "description": "Accumulated job warnings"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ],
            "description": "Failure details if status=failed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "JobStatus",
        "description": "Current lifecycle status and stage progression of a job."
      },
      "LayerFormat": {
        "type": "string",
        "enum": [
          "png",
          "geojson",
          "geoparquet",
          "cog"
        ],
        "title": "LayerFormat",
        "description": "File/wire format for layer."
      },
      "LayerMetadata": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Layer identifier e.g. flood_polygons, s1_post"
          },
          "layer_type": {
            "$ref": "#/components/schemas/LayerType",
            "description": "Raster or vector"
          },
          "format": {
            "$ref": "#/components/schemas/LayerFormat",
            "description": "Format encoding"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Endpoint URL to fetch layer data"
          },
          "bounds": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bounds",
            "description": "Layer bounding box [min_lon, min_lat, max_lon, max_lat]"
          },
          "feature_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Count",
            "description": "Feature count for vectors"
          },
          "legend": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legend",
            "description": "Legend and styling tokens"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "layer_type",
          "format",
          "url"
        ],
        "title": "LayerMetadata",
        "description": "Description and download/render URL of a generated layer."
      },
      "LayerType": {
        "type": "string",
        "enum": [
          "raster",
          "vector"
        ],
        "title": "LayerType",
        "description": "Type of map layer."
      },
      "OrbitDirection": {
        "type": "string",
        "enum": [
          "ASCENDING",
          "DESCENDING"
        ],
        "title": "OrbitDirection",
        "description": "Sentinel-1 orbit pass direction."
      },
      "PipelineStage": {
        "type": "string",
        "enum": [
          "pairing",
          "fetch",
          "flood_model",
          "terrain_filter",
          "osm",
          "damage",
          "isolation",
          "facts",
          "report"
        ],
        "title": "PipelineStage",
        "description": "Explicit pipeline stages defined in ARCHITECTURE.md Section 4."
      },
...
      "Provenance": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/ProvenanceMode",
            "description": "Execution mode: live, cached, or degraded"
          },
          "computed_at": {
            "type": "string",
            "title": "Computed At",
            "description": "ISO 8601 timestamp of computation"
          },
          "code_version": {
            "type": "string",
            "title": "Code Version",
            "description": "Git commit hash or release version"
          },
          "config_hash": {
            "type": "string",
            "title": "Config Hash",
            "description": "Hash of active configuration values"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version",
            "description": "Model weights identifier or baseline version"
          },
          "scenes": {
            "$ref": "#/components/schemas/ScenePairingProvenance",
            "description": "S1 pair details"
          },
          "s2": {
            "$ref": "#/components/schemas/S2Provenance",
            "description": "S2 optical details"
          },
          "osm_snapshot_date": {
            "type": "string",
            "title": "Osm Snapshot Date",
            "description": "Pre-event snapshot date used for OSM query"
          },
          "degradations": {
            "items": {
              "$ref": "#/components/schemas/DegradationItem"
            },
            "type": "array",
            "title": "Degradations",
            "description": "List of degradations if any"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "computed_at",
          "code_version",
          "config_hash",
          "model_version",
          "scenes",
          "s2",
          "osm_snapshot_date"
        ],
        "title": "Provenance",
        "description": "Complete provenance object attached to all output artifacts."
      },
      "ProvenanceMode": {
        "type": "string",
        "enum": [
          "live",
          "cached",
          "degraded"
        ],
        "title": "ProvenanceMode",
        "description": "Provenance execution mode."
      },
      "ResultManifest": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "bbox": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Bbox"
          },
          "event_date": {
            "type": "string",
            "title": "Event Date"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "layers": {
            "additionalProperties": {
              "$ref": "#/components/schemas/LayerMetadata"
            },
            "type": "object",
            "title": "Layers",
            "description": "Map of available raster and vector layers"
          },
          "facts_url": {
            "type": "string",
            "title": "Facts Url",
            "description": "URL to fetch full facts.json"
          },
          "report_url": {
            "type": "string",
            "title": "Report Url",
            "description": "URL to generate/fetch situation report PDF"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary",
            "description": "Headline metric summaries"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "bbox",
          "event_date",
          "provenance",
          "layers",
          "facts_url",
          "report_url"
        ],
        "title": "ResultManifest",
        "description": "Complete manifest returned upon job completion."
      },
      "S2Provenance": {
        "properties": {
          "used": {
            "type": "boolean",
            "title": "Used",
            "description": "Whether optical data was incorporated"
          },
          "cloud_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloud Pct",
            "description": "Estimated cloud percentage in AOI clip"
          },
          "scene_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Id",
            "description": "S2 L2A scene identifier if used"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "used"
        ],
        "title": "S2Provenance",
        "description": "Sentinel-2 optical layer usage and cloud coverage."
      },
      "SceneMetadata": {
        "properties": {
          "scene_id": {
            "type": "string",
            "title": "Scene Id",
            "description": "Product identifier / scene name"
          },
          "acquisition_time": {
            "type": "string",
            "title": "Acquisition Time",
            "description": "ISO 8601 acquisition timestamp"
          },
          "orbit_direction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrbitDirection"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass direction"
          },
          "relative_orbit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relative Orbit",
            "description": "Relative orbit number"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scene_id",
          "acquisition_time"
        ],
        "title": "SceneMetadata",
        "description": "Metadata for an individual satellite acquisition."
      },
      "ScenePairingProvenance": {
        "properties": {
          "post": {
            "$ref": "#/components/schemas/SceneMetadata",
            "description": "Post-event acquisition"
          },
          "pre": {
            "items": {
              "$ref": "#/components/schemas/SceneMetadata"
            },
            "type": "array",
            "title": "Pre",
            "description": "Pre-event baseline acquisition(s)"
          },
          "relative_orbit": {
            "type": "integer",
            "title": "Relative Orbit",
            "description": "Shared relative orbit number"
          },
          "direction": {
            "$ref": "#/components/schemas/OrbitDirection",
            "description": "Shared pass direction"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "post",
          "pre",
          "relative_orbit",
          "direction"
        ],
        "title": "ScenePairingProvenance",
        "description": "Verified same-orbit Sentinel-1 pair metadata."
      },
...
      "StageState": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "succeeded",
          "failed",
          "skipped"
        ],
        "title": "StageState",
        "description": "Execution status of an individual stage."
      },
      "StageStatus": {
        "properties": {
          "name": {
            "$ref": "#/components/schemas/PipelineStage",
            "description": "Stage identifier"
          },
          "state": {
            "$ref": "#/components/schemas/StageState",
            "description": "Execution status",
            "default": "pending"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At",
            "description": "ISO 8601 timestamp"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At",
            "description": "ISO 8601 timestamp"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds",
            "description": "Duration in seconds"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "description": "Non-fatal warnings recorded during stage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "StageStatus",
        "description": "Progress record for a single pipeline stage."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}